Added two arguments first and start for pagination.
first to collect the number of records that should be returned at once
in a single data page, and start to define the starting position or
index of the first record to return.

By default, the allUsers query returns only the first 50 users, and
the allPhotos query returns only the first 25 photos.